# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeechBloom API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  patient: 'patient',
  therapist: 'therapist',
  parent: 'parent',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  goals: string[];
  dailyMinutes: number;
  currentStreak: number;
  longestStreak: number;
  totalMinutes: number;
  onboardingComplete: boolean;
  createdAt: string;
}

export interface UserUpdate {
  name?: string;
  dailyMinutes?: number;
  goals?: string[];
}

export interface OnboardingInput {
  name: string;
  goals: string[];
  dailyMinutes: number;
}

export type ExerciseExerciseType = typeof ExerciseExerciseType[keyof typeof ExerciseExerciseType];


export const ExerciseExerciseType = {
  timer: 'timer',
  reps: 'reps',
  bilateral: 'bilateral',
} as const;

export interface Exercise {
  id: number;
  name: string;
  category: string;
  instructions: string;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  rounds?: number | null;
  exerciseType: ExerciseExerciseType;
  sortOrder: number;
}

export interface ExerciseLog {
  id: number;
  sessionId: number;
  exerciseId: number;
  exerciseName: string;
  /** @nullable */
  repsCompleted?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  completedAt: string;
}

export interface Session {
  id: number;
  userId: number;
  date: string;
  durationSeconds: number;
  exercisesCompleted: number;
  totalExercises: number;
  completionPercent: number;
  completed: boolean;
  logs?: ExerciseLog[];
}

export interface SessionInput {
  totalExercises: number;
}

export interface SessionUpdate {
  durationSeconds?: number;
  completed?: boolean;
}

export interface ExerciseLogInput {
  exerciseId: number;
  /** @nullable */
  repsCompleted?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
}

export interface ProgressSummary {
  totalSessions: number;
  currentStreak: number;
  longestStreak: number;
  totalMinutes: number;
  totalExercisesCompleted: number;
  todayCompletionPercent: number;
  weeklyGoalPercent: number;
  mostImprovedCategory?: string;
}

export interface DailyProgress {
  date: string;
  minutesPracticed: number;
  exercisesCompleted: number;
  sessionCount: number;
}

export interface HeatmapEntry {
  date: string;
  count: number;
}

export interface CategoryStat {
  category: string;
  exercisesCompleted: number;
  masteryPercent: number;
}

export interface Achievement {
  id: number;
  name: string;
  description: string;
  icon: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
}

export interface Patient {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  lastSession?: string | null;
  currentStreak: number;
  completionPercent: number;
  totalSessions: number;
}

export interface PatientInput {
  email: string;
  name: string;
}

export type ListExercisesParams = {
category?: string;
};