# lib/api-zod/src/generated/types/exerciseLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeechBloom API
 * OpenAPI spec version: 0.1.0
 */

export interface ExerciseLog {
  id: number;
  sessionId: number;
  exerciseId: number;
  exerciseName: string;
  /** @nullable */
  repsCompleted?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  completedAt: string;
}