# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeechBloom API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  lastSession?: string | null;
  currentStreak: number;
  completionPercent: number;
  totalSessions: number;
}