# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeechBloom API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressSummary {
  totalSessions: number;
  currentStreak: number;
  longestStreak: number;
  totalMinutes: number;
  totalExercisesCompleted: number;
  todayCompletionPercent: number;
  weeklyGoalPercent: number;
  mostImprovedCategory?: string;
}