# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpeechBloom API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  goals: string[];
  dailyMinutes: number;
  currentStreak: number;
  longestStreak: number;
  totalMinutes: number;
  onboardingComplete: boolean;
  createdAt: string;
}